requirements.txt
grpcio==1.57.0
numpy==1.24.1
protobuf==4.24.0
torch==2.0.1